refactoring(stream-client): Folder restructuring and naming changes of existing components of run-stream module  - Renamed run-stream-button to run-stream  - Renamed run-stream to simulation-configuration and moved to ./configuration folder  - Renamed action-status to simulation-controls and moved to ./simulation-controls folder  - Moved file-status component files into run-stream module
