--- libs/plugins/stream-client/src/lib/file-status/index.ts
@@ -1,1 +1,0 @@
-export * from './file-status';
