+++ libs/plugins/stream-client/src/lib/run-stream/configuration/simulation-configuration.component.less
