+++ libs/plugins/stream-client/src/lib/run-stream/file-status/file-status.component.ts
