+++ libs/plugins/stream-client/src/lib/run-stream/file-status/index.ts
@@ -1,0 +1,2 @@
+export * from './file-status';
+export * from './file-status.component';
