--- libs/plugins/stream-client/src/lib/run-stream/run-stream-button/run-stream-button.component.html
@@ -1,18 +1,0 @@
-<flogo-stream-run-action-status
-  [status]="simulatorStatus$ | async"
-  [disableRunStream]="disableRunStream"
-  (run)="runStream()"
-  (pause)="pauseSimulation()"
-  (stop)="stopSimulation()"
-  (resume)="resumeSimulation()"
-></flogo-stream-run-action-status>
-
-<flogo-stream-run-stream
-  *ngIf="showFileInput"
-  [resourceId]="resourceId"
-  [fileName]="fileName"
-  [fileUploadStatus]="fileUploadStatus"
-  [mappingType]="simulationConfig?.inputMappingType"
-  (setFilePath)="setFilePath($event)"
-  (startSimulation)="startSimulation($event)"
-></flogo-stream-run-stream>
