+++ libs/plugins/stream-client/src/lib/run-stream/run-stream.component.html
@@ -1,60 +1,18 @@
-<div class="import-header">
-  Import stream inputs from
-</div>
-<div
-  class="file-upload"
-  [class.file-upload--empty]="fileUploadStatus !== FileStatus.Uploaded"
-  fgDragDropFile
-  [status]="fileUploadStatus"
-  (fileDropped)="uploadFile($event)"
-  (dragging)="setIsDragging($event)"
->
-  <flogo-stream-file-status
-    [status]="fileUploadStatus"
-    [isDragging]="isDragging"
-    (remove)="removeFile()"
-  ></flogo-stream-file-status>
+<flogo-stream-simulation-controls
+  [status]="simulatorStatus$ | async"
+  [disableRunStream]="disableRunStream"
+  (run)="runStream()"
+  (pause)="pauseSimulation()"
+  (stop)="stopSimulation()"
+  (resume)="resumeSimulation()"
+></flogo-stream-simulation-controls>
 
-  <div
-    class="upload-instructions"
-    *ngIf="fileUploadStatus !== FileStatus.Uploaded; else uploadedState"
-  >
-    <span class="upload-instructions__lead">
-      Drag and drop
-    </span>
-    <span>
-      your file here, or
-    </span>
-    <span class="upload-instructions__link" (click)="fileInput.click()">
-      browse to upload
-    </span>
-    <input type="file" #fileInput hidden (change)="uploadFile($event.target.files)" />
-  </div>
-
-  <ng-template #uploadedState>
-    <div class="file-name">{{ fileName }}</div>
-    <flogo-select
-      [value]="mappingTypeSelection"
-      label="{{ 'STREAMS.SIMULATOR:MAPPING-LABEL' | translate }}"
-      (select)="onChangeMappingType($event)"
-    >
-      <flogo-select-option
-        [value]="InputMappingType.SingleInput"
-        label="{{ 'STREAMS.SIMULATOR:MAPPING-OPTION-SINGLE' | translate }}"
-      ></flogo-select-option>
-      <flogo-select-option
-        [value]="InputMappingType.SeparateByColumn"
-        label="{{ 'STREAMS.SIMULATOR:MAPPING-OPTION-SEPARATE' | translate }}"
-      ></flogo-select-option>
-    </flogo-select>
-  </ng-template>
-</div>
-<div class="footer">
-  <button
-    [disabled]="disableRunStream"
-    (click)="runSimulation()"
-    class="flogo-button--default run-stream"
-  >
-    {{ 'STREAMS.SIMULATOR:RUN-STREAM' | translate }}
-  </button>
-</div>
+<flogo-stream-simulation-configuration
+  *ngIf="showFileInput"
+  [resourceId]="resourceId"
+  [fileName]="fileName"
+  [fileUploadStatus]="fileUploadStatus"
+  [mappingType]="simulationConfig?.inputMappingType"
+  (setFilePath)="setFilePath($event)"
+  (startSimulation)="startSimulation($event)"
+></flogo-stream-simulation-configuration>
