+++ libs/plugins/stream-client/src/lib/run-stream/simulation-controls/index.ts
@@ -1,0 +1,1 @@
+export { SimulationControlsComponent } from './simulation-controls.component';
