+++ libs/plugins/stream-client/src/lib/run-stream/simulation-controls/simulation-controls.component.html
